+++ tests/data/generate_source_of_truth.py
@@ -96,6 +96,7 @@ def generate_indicators():
             {"kind": "aroon"},
             {"kind": "donchian"},
             {"kind": "cmo"},
+            {"kind": "mfi"},
         ],
     )
 
@@ -124,6 +125,7 @@ def generate_indicators():
     save_json_result([round_values(value) for value in df["STDEV_30"].tolist()], "STDEV")
     save_json_result([round_values(value) for value in df["TSI_13_25_13"].tolist()], "TSI")
     save_json_result([round_values(value) for value in df["CMO_14"].tolist()], "CMO")
+    save_json_result([round_values(value) for value in df["MFI_14"].tolist()], "MFI")
 
     kc_data = []
     for kc in zip(