+++ CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.1.3
+
 - Added Indicators
   - Chande Momentum Oscillator (CMO)
   - High Low Close Average (HLCA) 
+  - Money Flow Index (MFI)
+
 
 ## 1.1.2 - 2024-05-03